Clean up classes a bit more
