+++ Morphic.Server/Email/EmailJob.cs
@@ -170,25 +170,12 @@ namespace Morphic.Server.Email
         private class UnableToSendEmailException : SendEmailException
         {
         }
-    }
-
-    class EmailJobException : MorphicServerException
-    {
-        public EmailJobException(string error) : base(error)
+        
+        protected class EmailJobException : MorphicServerException
         {
-        }
-    }
-
-    public static class EmailConstants
-    {
-        public enum EmailTypes
-        {
-            None = 0,
-            WelcomeEmailValidation = 1,
-            PasswordReset = 2,
-            PasswordResetEmailNotValidated = 3,
-            PasswordResetUnknownEmail = 4,
-            ChangePasswordEmail = 5
+            public EmailJobException(string error) : base(error)
+            {
+            }
         }
     }
 }
